requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
passlib[bcrypt]
pydantic[email]
pydantic>=2.0
pydantic-settings>=2.0
pydantic-core>=2.0.0